+++ apps/router-e2e/__e2e__/server/app/api/problematic+api.ts
@@ -1,0 +1,5 @@
+import { ExpoResponse } from 'expo-router/server';
+
+export function GET(): ExpoResponse {
+  throw new Error('This is a test error');
+}