+++ apps/router-e2e/app.config.js
@@ -28,7 +28,7 @@ module.exports = {
     typedRoutes: true,
   },
   web: {
-    output: process.env.EXPO_USE_STATIC === '1' ? 'static' : 'single',
+    output: process.env.EXPO_USE_STATIC ?? 'static',
     bundler: 'metro',
   },
   plugins: [