+++ packages/@expo/cli/src/export/createMetadataJson.ts
@@ -28,25 +28,29 @@ export function createMetadataJson({
     version: 0,
     bundler: 'metro',
     fileMetadata: Object.entries(bundles).reduce<Record<string, Partial<PlatformMetadata>>>(
-      (metadata, [platform, bundle]) => ({
-        ...metadata,
-        [platform]: {
-          // Get the filename for each platform's bundle.
-          bundle: path.join('bundles', fileNames[platform]!),
-          // Collect all of the assets and convert them to the serial format.
-          assets: bundle.assets
-            .map(
-              (asset) =>
-                // Each asset has multiple hashes which we convert and then flatten.
-                asset.fileHashes?.map((hash) => ({
-                  path: path.join('assets', hash),
-                  ext: asset.type,
-                }))
-            )
-            .filter(Boolean)
-            .flat(),
-        },
-      }),
+      (metadata, [platform, bundle]) => {
+        if (platform === 'web') return metadata;
+
+        return {
+          ...metadata,
+          [platform]: {
+            // Get the filename for each platform's bundle.
+            bundle: path.join('bundles', fileNames[platform]!),
+            // Collect all of the assets and convert them to the serial format.
+            assets: bundle.assets
+              .map(
+                (asset) =>
+                  // Each asset has multiple hashes which we convert and then flatten.
+                  asset.fileHashes?.map((hash) => ({
+                    path: path.join('assets', hash),
+                    ext: asset.type,
+                  }))
+              )
+              .filter(Boolean)
+              .flat(),
+          },
+        };
+      },
       {}
     ) as FileMetadata,
   };