+++ packages/@expo/cli/src/export/writeContents.ts
@@ -38,11 +38,11 @@ function createBundleHash(bundle: string | Uint8Array): string {
 export async function writeBundlesAsync({
   bundles,
   outputDir,
-  useWebSSG,
+  useServerRendering,
 }: {
   bundles: Partial<Record<Platform, Pick<BundleOutput, 'hermesBytecodeBundle' | 'code'>>>;
   outputDir: string;
-  useWebSSG?: boolean;
+  useServerRendering?: boolean;
 }) {
   const hashes: Partial<Record<Platform, string>> = {};
   const fileNames: Partial<Record<Platform, string>> = {};
@@ -52,7 +52,7 @@ export async function writeBundlesAsync({
     Pick<BundleOutput, 'hermesBytecodeBundle' | 'code'>,
   ][]) {
     // TODO: Move native to use the newer `_expo/...` bundle writing system.
-    if (platform === 'web' && useWebSSG) {
+    if (platform === 'web' && useServerRendering) {
       continue;
     }
     const bundle = bundleOutput.hermesBytecodeBundle ?? bundleOutput.code;