+++ packages/@expo/cli/src/start/server/metro/router.ts
@@ -1,4 +1,6 @@
+import { ExpoConfig } from '@expo/config';
 import chalk from 'chalk';
+import { sync as globSync } from 'glob';
 import path from 'path';
 import resolveFrom from 'resolve-from';
 
@@ -37,6 +39,17 @@ function getFallbackEntryRoot(projectRoot: string): string {
   return path.join(projectRoot, 'node_modules/expo-router/entry');
 }
 
+export function getRouterDirectoryModuleIdWithManifest(
+  projectRoot: string,
+  exp: ExpoConfig
+): string {
+  return exp.extra?.router?.unstable_src ?? getRouterDirectory(projectRoot);
+}
+
+export function getRouterDirectoryWithManifest(projectRoot: string, exp: ExpoConfig): string {
+  return path.join(projectRoot, getRouterDirectoryModuleIdWithManifest(projectRoot, exp));
+}
+
 export function getRouterDirectory(projectRoot: string): string {
   // more specific directories first
   if (directoryExistsSync(path.join(projectRoot, 'src/app'))) {
@@ -46,4 +59,26 @@ export function getRouterDirectory(projectRoot: string): string {
 
   Log.debug('Using app as the root directory for Expo Router.');
   return 'app';
+}
+
+export function isApiRouteConvention(name: string): boolean {
+  return /\+api\.[tj]sx?$/.test(name);
+}
+
+export function getApiRoutesForDirectory(cwd: string) {
+  return globSync('**/*+api.@(ts|tsx|js|jsx)', {
+    cwd,
+    absolute: true,
+  });
+}
+
+// Used to emulate a context module, but way faster. TODO: May need to adjust the extensions to stay in sync with Metro.
+export function getRoutePaths(cwd: string) {
+  return globSync('**/*.@(ts|tsx|js|jsx)', {
+    cwd,
+  }).map((p) => './' + normalizePaths(p));
+}
+
+function normalizePaths(p: string) {
+  return p.replace(/\\/g, '/');
 }